frontend: Rename HomeView to BoardsView and BoardView to NotesView
