+++ frontend/src/router/index.ts
@@ -1,21 +1,19 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import HomeView from '@/views/HomeView.vue'
+import BoardsView from '@/views/BoardsView.vue'
+import NotesView from '@/views/NotesView.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     {
       path: '/',
-      name: 'home',
-      component: HomeView,
+      name: 'boards',
+      component: BoardsView,
     },
     {
       path: '/board/:id',
       name: 'board',
-      // route level code-splitting
-      // this generates a separate chunk (About.[hash].js) for this route
-      // which is lazy-loaded when the route is visited.
-      component: () => import('../views/BoardView.vue'),
+      component: NotesView,
     },
   ],
 })
