+++ frontend/src/views/NotesView.vue
